mock in the cli
