+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.sln
@@ -15,8 +15,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Sdk.Tools.Cli.Benchma
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Sdk.Tools.CodeownersUtils", "..\codeowners-utils\Azure.Sdk.Tools.CodeownersUtils\Azure.Sdk.Tools.CodeownersUtils.csproj", "{D27C2C44-3AC4-0732-FF87-DD1697A9DF37}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Azure.Sdk.Tools.Mock", "Azure.Sdk.Tools.Mock\Azure.Sdk.Tools.Mock.csproj", "{F6CA0C60-A387-4322-A26C-0980EE8C1358}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -99,18 +97,6 @@ Global
 		{D27C2C44-3AC4-0732-FF87-DD1697A9DF37}.Release|x64.Build.0 = Release|Any CPU
 		{D27C2C44-3AC4-0732-FF87-DD1697A9DF37}.Release|x86.ActiveCfg = Release|Any CPU
 		{D27C2C44-3AC4-0732-FF87-DD1697A9DF37}.Release|x86.Build.0 = Release|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Debug|x64.Build.0 = Debug|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Debug|x86.Build.0 = Debug|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Release|x64.ActiveCfg = Release|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Release|x64.Build.0 = Release|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Release|x86.ActiveCfg = Release|Any CPU
-		{F6CA0C60-A387-4322-A26C-0980EE8C1358}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
