--- tools/azsdk-cli/Azure.Sdk.Tools.Mock/Azure.Sdk.Tools.Mock.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <AssemblyName>azsdk-mock</AssemblyName>
-    <NoWarn>ASP0000</NoWarn>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <StaticWebAssetsEnabled>false</StaticWebAssetsEnabled>
-    <IsWebConfigTransformDisabled>true</IsWebConfigTransformDisabled>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="ModelContextProtocol" Version="0.5.0-preview.1" />
-    <PackageReference Include="ModelContextProtocol.AspNetCore" Version="0.5.0-preview.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../Azure.Sdk.Tools.Cli/Azure.Sdk.Tools.Cli.csproj" />
-  </ItemGroup>
-
-</Project>
