--- tools/azsdk-cli/Azure.Sdk.Tools.Mock/Program.cs
@@ -1,25 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using Azure.Sdk.Tools.Mock;
-using Azure.Sdk.Tools.Mock.Handlers;
-
-var builder = WebApplication.CreateBuilder(args);
-
-// Suppress noisy framework logging — mock server only needs minimal output
-builder.Logging.SetMinimumLevel(LogLevel.Warning);
-
-// Register the mock tool factory (auto-discovers IMockToolHandler implementations)
-builder.Services.AddSingleton<MockToolFactory>();
-
-// Register mock MCP tools (same names/schemas as real CLI, but with mock responses)
-MockToolRegistrations.RegisterMockMcpTools(builder.Services);
-
-// Set up MCP server with stdio transport (same as the real CLI)
-builder.Services
-    .AddMcpServer()
-    .WithStdioServerTransport();
-
-var app = builder.Build();
-
-await app.RunAsync();
